[ApartmentService] Response status code implementation

+++ Apartments/src/services/ApartmentService.java
@@ -94,10 +94,6 @@ public class ApartmentService {
 	@Path("/getMyApartments")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response getJustMyApartments() {
-		//isUserHost();
-		// With this, we get user who is logged in.
-		// We are in UserService method login() tie user for session.
-		// And now we can get him.
 		if(isUserHost()) {
 			User user = (User) request.getSession().getAttribute("loginUser");
 			ApartmentsDAO apartmentsDAO = getApartments();
@@ -160,14 +156,12 @@ public class ApartmentService {
 
 		if(isUserHost()) {
 			User user = (User) request.getSession().getAttribute("loginUser");
-			System.out.println("\n\n\t\tSTIGAO JE APARTMAN SA ID-om: " + newItem.addedApartment.getID() + " i cenom"
-					+ newItem.addedApartment.getPricePerNight() + "\n\n");
 		
 			ApartmentsDAO apartmentsDAO = getApartments();
 			apartmentsDAO.changeApartment(newItem.addedApartment, newItem.startDateForReservation, newItem.endDateForReservation);
 			
 			return Response
-					.status(Response.Status.ACCEPTED).entity("SUCCESS ACTIVATED")
+					.status(Response.Status.ACCEPTED).entity("SUCCESS CHANGED")
 					.entity(apartmentsDAO.getHostApartments(user))
 					.build();
 		}	
@@ -201,8 +195,49 @@ public class ApartmentService {
 				}	
 			}
 			return Response
-					.status(Response.Status.ACCEPTED).entity("SUCCESS ACTIVATED")
+					.status(Response.Status.ACCEPTED).entity("SUCCESS")
 					.entity(lsita)
+					.build();
+		}
+		return Response.status(403).type("text/plain")
+				.entity("You do not have permission to access!").build();
+	}
+
+	@POST
+	@Path("/changeApartment")
+	@Consumes(MediaType.APPLICATION_JSON)
+	@Produces(MediaType.APPLICATION_JSON)
+	public Response changeApartment(ApartmentDTOJSON updatedApartment) {
+		
+		if(isUserAdmin() || isUserHost()) {
+			ApartmentsDAO apartments = getApartments();
+			apartments.changeApartment(updatedApartment.addedApartment, updatedApartment.startDateForReservation, updatedApartment.endDateForReservation);
+	
+			return Response
+					.status(Response.Status.ACCEPTED).entity("SUCCESS CHANGED")
+					.entity(getApartments().getValues())
+					.build();
+		}
+		return Response.status(403).type("text/plain")
+				.entity("You do not have permission to access!").build();
+	}
+
+	@DELETE
+	@Path("/deleteHostApartment")
+	@Consumes(MediaType.APPLICATION_JSON)
+	@Produces(MediaType.APPLICATION_JSON)
+	public Response deleteHostApartment(ApartmentsDTO updatedApartment) {
+		if(isUserHost()) {
+
+			ApartmentsDAO apartmentsDAO = getApartments();
+			apartmentsDAO.deleteApartment(updatedApartment.identificator);
+	
+			User user = (User) request.getSession().getAttribute("loginUser");
+			UsersDAO allUsersDAO = getUsers();
+			allUsersDAO.deleteHostApartment(user.getID(), updatedApartment.identificator);
+			return Response
+					.status(Response.Status.ACCEPTED).entity("SUCCESS CHANGED")
+					.entity(apartmentsDAO.getHostApartments(user))
 					.build();
 		}
 		return Response.status(403).type("text/plain")
@@ -210,55 +245,25 @@ public class ApartmentService {
 		
 	}
 
-	@POST
-	@Path("/changeApartment")
-	@Consumes(MediaType.APPLICATION_JSON)
-	@Produces(MediaType.APPLICATION_JSON)
-	public Collection<Apartment> changeApartment(ApartmentDTOJSON updatedApartment) {
-		System.out.println("\n\n\t\tSTIGAO JE APARTMAN SA ID-om: " + updatedApartment.addedApartment.getID() + " i cenom"
-				+ updatedApartment.addedApartment.getPricePerNight() + "\n\n");
-
-		ApartmentsDAO apartments = getApartments();
-		apartments.changeApartment(updatedApartment.addedApartment, updatedApartment.startDateForReservation, updatedApartment.endDateForReservation);
-
-		return getApartments().getValues();
-	}
-
-	@DELETE
-	@Path("/deleteHostApartment")
-	@Consumes(MediaType.APPLICATION_JSON)
-	@Produces(MediaType.APPLICATION_JSON)
-	public Collection<Apartment> deleteHostApartment(ApartmentsDTO updatedApartment) {
-		System.out.println("\n\n\t\tSTIGAO JE APARTMAN SA ID-om: " + updatedApartment.identificator + "\n\n");
-
-		ApartmentsDAO apartmentsDAO = getApartments();
-		apartmentsDAO.deleteApartment(updatedApartment.identificator);
-
-		// With this, we get user who is logged in.
-		// We are in UserService method login() tie user for session.
-		// And now we can get him.
-		User user = (User) request.getSession().getAttribute("loginUser");
-		UsersDAO allUsersDAO = getUsers();
-		allUsersDAO.deleteHostApartment(user.getID(), updatedApartment.identificator);
-		
-
-		return apartmentsDAO.getHostApartments(user);
-	}
-
 	@DELETE
 	@Path("/deleteApartment")
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
-	public Collection<Apartment> deleteApartment(ApartmentsDTO updatedApartment) {
-		System.out.println("\n\n\t\tSTIGAO JE APARTMAN SA ID-om: " + updatedApartment.identificator + "\n\n");
-
-		ApartmentsDAO apartments = getApartments();
-		apartments.deleteApartment(updatedApartment.identificator);
-		
-		UsersDAO allUsersDAO = getUsers();
-		allUsersDAO.deleteHostApartment(updatedApartment.hostID, updatedApartment.identificator);
-
-		return getApartments().getValues();
+	public Response deleteApartment(ApartmentsDTO updatedApartment) {
+		if(isUserAdmin()) {
+			ApartmentsDAO apartments = getApartments();
+			apartments.deleteApartment(updatedApartment.identificator);
+			
+			UsersDAO allUsersDAO = getUsers();
+			allUsersDAO.deleteHostApartment(updatedApartment.hostID, updatedApartment.identificator);
+	
+			return Response
+					.status(Response.Status.ACCEPTED).entity("SUCCESS CHANGED")
+					.entity(getApartments().getValues())
+					.build();
+		}
+		return Response.status(403).type("text/plain")
+				.entity("You do not have permission to access!").build();
 	}
 
 	private ApartmentsDAO getApartments() {
